make sectionRange a float,

+++ D3D11Engine/D3D11GraphicsEngine.cpp
@@ -4092,7 +4092,7 @@ void XM_CALLCONV D3D11GraphicsEngine::DrawWorldAround_Layered(
 
 /** Draws everything around the given position */
 void XM_CALLCONV D3D11GraphicsEngine::DrawWorldAround( FXMVECTOR position,
-    int sectionRange, float vobXZRange,
+    float sectionRange, float vobXZRange,
     bool cullFront, bool dontCull ) {
     // Setup renderstates
     Engine::GAPI->GetRendererState().RasterizerState.SetDefault();
